Added rudimentary code for gain
Added gain function to the code. So we can control output file gain based on user input. HOWEVER for now to avoid bugs, I've added this as a seperate loop in the code after the main loop.
This reduces complexity but adds a lot of overhead processing the file twice. 
After we have the main loop working we can incorporate it into the main loop